monitor/requirements.txt
yfinance>=0.2.36,<0.3
pandas>=2.1,<3
numpy>=1.26,<2
pyarrow>=14.0
python-dotenv>=1.0
